move DagTask to SniperKernel
