--- Examples/DepGraphUsages/python/DepGraphUsages.py
@@ -1,4 +1,0 @@
-import Sniper
-Sniper.loadDll("libDepGraphUsages.so")
-del Sniper
-
