+++ SniperKernel/SniperPrivate/AlgNode.h
