+++ SniperPython/src/ModuleDef.cc
@@ -1,17 +1,17 @@
 /* Copyright (C) 2018-2021
    Institute of High Energy Physics and Shandong University
    This file is part of SNiPER.
- 
+
    SNiPER is free software: you can redistribute it and/or modify
    it under the terms of the GNU Lesser General Public License as published by
    the Free Software Foundation, either version 3 of the License, or
    (at your option) any later version.
- 
+
    SNiPER is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU Lesser General Public License for more details.
- 
+
    You should have received a copy of the GNU Lesser General Public License
    along with SNiPER.  If not, see <http://www.gnu.org/licenses/>. */
 
@@ -43,6 +43,7 @@ void export_Sniper_SniperLog();
 void export_SniperPython_PyDataStore();
 void export_Sniper_Timer();
 void export_Sniper_TimerSvc();
+void export_Sniper_DagTask();
 
 BOOST_PYTHON_MODULE(libSniperPython)
 {
@@ -66,4 +67,5 @@ BOOST_PYTHON_MODULE(libSniperPython)
     export_SniperPython_PyDataStore();
     export_Sniper_Timer();
     export_Sniper_TimerSvc();
+    export_Sniper_DagTask();
 }
