--- SniperUtil/DagTask/python/DagTask.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-# -*- coding:utf-8 -*-
-
-import Sniper
-Sniper.loadDll("libDagTaskLib.so")
-del Sniper
