--- SniperUtil/DagTask/python/DagTask/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-import sys
-sys.setdlopenflags( 0x100 | 0x2 ) # RTLD_GLOBAL | RTLD_NOW
-
-from .libDagTask import *
